# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Authentication App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /user/{id}:
    x-swagger-router-controller: user
    get:
      operationId: getOne
      description: get a user
      
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
     
    put:
      operationId: update
      description: update a user
      
      parameters:
        - name: id
          description: User id
          type: number
          in: path
          required: true
        - name: username
          description: user name
          in: body
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
     
    delete:
      operationId: delUser
      description: delete a user
      # define the parameters
      parameters:
        - name: id
          description: User id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  GetUserResponse:
    required:
      - id
      - username
      - email
    properties:
      id:
        type: number
      username: 
        type: string
      email:
        type: string
  User:
    type: object
    properties:
      username:
        type: string
        description: user name
      email:
        type: string
        description: email address
    required:
      - username
      - email
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description